requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Code quality
black==23.12.1
flake8==7.0.0
pylint==3.0.3

# Security scanning
bandit==1.7.6
safety==2.3.5
